Cleanup, derived pipeline example, bug fixes
